Fixed dexterity referenceablebehavior integration.
Breaches would be correctly found, but the code would give an error
when trying to get the Title of the source, which in the dexterity
case was an adapter.

We could add the Title on the adapter, but more changes would be
needed, like also adding absolute_url and changing the isAccessible
method to check the 'Access contents information' not on the adapter
but on its context.

It seemed better to simply not let store the actual dexterity object
as the source, instead of storing the adapter.